+++ client/src/App.js
@@ -8,7 +8,8 @@ import Main from "./pages/Main";
 import MyPage from "./pages/Mypage";
 import Signup from "./pages/Signup";
 import ArticleDetail from "./pages/detail/ArticleDetail";
-import SubCategory from "./components/subcategories/SubCategory";
+import SubCategory from "./pages/SubCategory";
+import ShoppingCart from './pages/ShoppingCart'
 
 const MainContainter = styled.div`
   /* display: flex;
@@ -29,6 +30,7 @@ function App() {
             <Route path="/signup" element={<Signup />} />
             <Route path="/detail/:id" element={<ArticleDetail />} />
             <Route path="/sub" element={<SubCategory />} />
+            <Route path="/cart" element={<ShoppingCart />} />
           </Routes>
         </MainContainter>
         <Footer />
