+++ client/src/pages/ShoppingCart.js
@@ -1,0 +1,15 @@
+import styled from "styled-components/macro";
+
+const CartBlock = styled.div`
+     margin: 30px 40px;
+`; 
+
+function ShoppingCart() {
+  return(
+    <CartBlock>
+
+    </CartBlock>      
+  )
+}
+
+export default ShoppingCart;
